requirements.txt
Django==1.8.6
djangorestframework==3.3.1
wheel==0.24.0
django-filter==0.11.0
django-crispy-forms==1.5.2